requirements.txt
# --- Core ---
streamlit==1.51.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
PyMuPDF==1.26.6
PyYAML>=6.0

# --- Vector DB ---
chromadb==0.5.15
pydantic>=2.8.2

# --- Embeddings stack (CPU) ---
sentence-transformers==2.6.1
transformers==4.44.2
tokenizers==0.19.1
safetensors==0.4.4
torch==2.5.1

# --- Database clients ---
psycopg[binary]>=3.1.0
